requirements.txt
APScheduler==3.11.0
cloudscraper==1.2.71
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
line-bot-sdk==3.19.0